requirements.txt
dht11
ISStreamer
python-dotenv
RPi.GPIO